Disambiguate pow()
svn path=/trunk/KDE/kdebase/workspace/; revision=746989

+++ gui/SensorDisplayLib/SignalPlotter.cc
@@ -608,7 +608,7 @@ void KSignalPlotter::calculateNiceRange()
     // Massage the range so that the grid shows some nice values.
     double step = mNiceRange / (mScaleDownBy*(mHorizontalLinesCount+1));
     int logdim = (int)floor( log10( step ) );
-    double dim = pow( 10, logdim ) / 2;
+    double dim = pow( (double)10.0, logdim ) / 2;
     int a = (int)ceil( step / dim );
     if(logdim >= 0)
         mPrecision = 0;
